refactor: 외부 api 트랜잭션 외부로 분리
